flwr_docker/test/pyproject.toml
# =====================================================================
# For a full TOML configuration guide, check the Flower docs:
# https://flower.ai/docs/framework/how-to-configure-pyproject-toml.html
# =====================================================================

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "test"
version = "1.0.0"
description = ""
license = "Apache-2.0"
# Dependencies for your Flower App
dependencies = [
    "flwr[simulation]>=1.22.0",
    "flwr-datasets[vision]>=0.5.0",
    "torch==2.8.0",
    "torchvision==0.23.0",
    "scikit-learn==1.7.2",
    "onnx==1.19.1",
    "onnx2torch==1.5.15",
    "protobuf==6.33.0",
    "grpcio==1.76.0",
    "grpcio-tools==1.76.0",
    "onnxscript==0.5.4"
]

[tool.hatch.build]
exclude = [
    "datasets/*", # Exclude datasets from FAB (if generated in this directory)
    "_static/*",  # Exclude images in README from FAB
    "data/*",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "pi"

# Point to your ServerApp and ClientApp objects
# Format: "<module>:<object>"
[tool.flwr.app.components]
# serverapp = "test_onnx.server_side.server_app:app"
# clientapp = "test_onnx.server_side.client_app:app"
serverapp = "test.server_app:app"
clientapp = "test.client_app:app"

# Custom config values accessible via `context.run_config`
[tool.flwr.app.config]
num-server-rounds = 2
fraction-evaluate = 0.5
local-epochs = 1
learning-rate = 0.01
batch-size = 64

# Default federation to use when running the app
[tool.flwr.federations]
default = "remote-federation"

# Local simulation federation with 10 virtual SuperNodes
[tool.flwr.federations.local-simulation]
options.num-supernodes = 10

# Remote federation example for use with SuperLink
[tool.flwr.federations.remote-federation]
address = "127.0.0.1:9093"
insecure = true            # Remove this line to enable TLS
# root-certificates = "<PATH/TO/ca.crt>"  # For TLS setup
